+++ src/MarkdownRenderer/multiLineElementLexer.js
@@ -51,6 +51,7 @@ const parseMultiLineElements = (list, supportedFeatures) => {
 
   // handle children elements recursively:
   result = result.map(l => {
+    // TODO Can we move this later (like we do for single-line) and avoid the `!== quotation` check?
     if (l.children && l.type !== 'quotation') {
       return {
         ...l,