+++ src/References/index.js
@@ -1,0 +1,5 @@
+import Citation from './Citation';
+import Reference from './Reference';
+
+export default { Citation, Reference };
+export { Citation, Reference };